backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.6.0
redis
pydantic>=2.6.0
pydantic-settings
python-dotenv
numpy
pandas
scikit-learn
scipy
sentence-transformers
faiss-cpu
openai
aiofiles
python-multipart
requests
openpyxl
dnspython
